Move `Tileset` to Raylib folder
