+++ src/Objects/Character.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Graphics/Tileset.h"
+#include "Raylib/Tileset.h"
 #include "Objects/Items/Item.h"
 #include "Objects/Types/Collidable.h"
 #include "Tiled/Object.h"
@@ -38,7 +38,7 @@ namespace Redge
 		auto GetHitBox() const -> Rectangle;
 		auto GetTextureOffset() const -> Vector2;
 
-		Tileset m_Animations = Tileset("assets/Entities/Character.png", 5, 4);
+		Raylib::Tileset m_Animations = Raylib::Tileset("assets/Entities/Character.png", 5, 4);
 		Animation m_Animation = Animation::Down;
 
 		uint16_t m_CurrentFrame = 0;
@@ -52,19 +52,19 @@ namespace Redge
 		float m_CharacterSpeed;
 		float m_SpeedMultiplier = 1;
 
-		Tileset m_HealthBar = Tileset("assets/UI/HealthBar.png", 1, 3);
+		Raylib::Tileset m_HealthBar = Raylib::Tileset("assets/UI/HealthBar.png", 1, 3);
 		float m_MaxHealth;
 		float m_Health;
 		float m_MaxOxygen;
 		float m_Oxygen;
 
-		Tileset m_CrystalIcon = Tileset("assets/UI/CrystalIcon.png", 1, 1);
+		Raylib::Tileset m_CrystalIcon = Raylib::Tileset("assets/UI/CrystalIcon.png", 1, 1);
 		uint32_t m_CrystalCount = 0;
 
-		Tileset m_InventoryIcon = Tileset("assets/UI/InventoryItem.png", 1, 1);
+		Raylib::Tileset m_InventoryIcon = Raylib::Tileset("assets/UI/InventoryItem.png", 1, 1);
 		std::array<std::unique_ptr<Item>, 5> m_Items{};
 
-		Tileset m_WeaponSlots = Tileset("assets/UI/WeaponSlots.png", 1, 1);
+		Raylib::Tileset m_WeaponSlots = Raylib::Tileset("assets/UI/WeaponSlots.png", 1, 1);
 		bool m_PrimaryWeaponSelected = false;
 	};
 } // namespace Redge
