+++ src/Objects/Torch.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Tiled/Object.h"
-#include "Graphics/Tileset.h"
+#include "Raylib/Tileset.h"
 
 #include <array>
 
@@ -21,7 +21,7 @@ namespace Redge
 
 	private:
 		Vector2 position;
-		Tileset m_Animation = Tileset("assets/Tilesets/objects/torch.png", 6, 1);
+		Raylib::Tileset m_Animation = Raylib::Tileset("assets/Tilesets/objects/torch.png", 6, 1);
 		uint16_t m_AnimationFrame = 0;
 		static constexpr float s_FrameDuration = 0.2;
 		float m_Frametime = 0;
