+++ src/Objects/UpgradeStation.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Graphics/Tileset.h"
+#include "Raylib/Tileset.h"
 #include "Objects/Types/Collidable.h"
 #include "Tiled/Object.h"
 
@@ -33,8 +33,8 @@ namespace Redge
 		static constexpr float s_FrameDuration = 0.2;
 		float m_FrameTime = 0;
 
-		Tileset m_Column = Tileset("assets/Tilesets/objects/upgradeStand.png", 1, 1);
-		Tileset m_Ghosts = Tileset("assets/Tilesets/objects/upgradeGhosts.png", 4, 5);
+		Raylib::Tileset m_Column = Raylib::Tileset("assets/Tilesets/objects/upgradeStand.png", 1, 1);
+		Raylib::Tileset m_Ghosts = Raylib::Tileset("assets/Tilesets/objects/upgradeGhosts.png", 4, 5);
 		uint8_t m_AnimationFrame = 0;
 		uint8_t m_AnimationColor;
 	};
